+++ cbits/bytes.c
@@ -1,6 +1,7 @@
 /*
 Copyright (c) 2017-2019 Dong Han
 Copyright Johan Tibell 2011, Dong Han 2019
+Copyright Dmitry Ivanov 2020
 All rights reserved.
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are met:
@@ -30,6 +31,11 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <turbob64.h>
 #ifdef __AVX2__
 #include <turbob64avx2.c>
+#endif
+
+#if defined(__x86_64__)
+#include <emmintrin.h>
+#include <xmmintrin.h>
 #endif
 
 HsInt hs_memchr(uint8_t *a, HsInt aoff, uint8_t b, HsInt n) {
@@ -267,4 +273,37 @@ HsInt hs_base64_decode(uint8_t* output, const uint8_t* input, HsInt off, HsInt l
 #else
     return (HsInt)tb64xdec(input+off, (size_t)len, output);
 #endif
+}
+
+/*  duplicate a string, interspersing the character through the elements
+    of the duplicated string
+    from https://github.com/haskell/bytestring/pull/310/files
+*/
+void hs_intersperse(unsigned char *q, 
+                     unsigned char *p, HsInt p_offset,
+                     HsInt n,
+                     unsigned char c) {
+#if defined(__x86_64__)
+  {
+    const __m128i separator = _mm_set1_epi8(c);
+    p += p_offset;
+    const unsigned char *const p_begin = p;
+    const unsigned char *const p_end = p_begin + n - 9;
+    while (p < p_end) {
+      const __m128i eight_src_bytes = _mm_loadl_epi64((__m128i *)p);
+      const __m128i sixteen_dst_bytes = _mm_unpacklo_epi8(eight_src_bytes, separator);
+      _mm_storeu_si128((__m128i *)q, sixteen_dst_bytes);
+      p += 8;
+      q += 16;
+    }
+    n -= p - p_begin;
+  }
+#endif
+    while (n > 1) {
+        *q++ = *p++;
+        *q++ = c;
+        n--;
+    }
+    if (n == 1)
+        *q = *p;
 }